refactor: recruitment 내 도메인 패키지 위치 변경
