+++ src/main/java/com/unionmate/backend/domain/applicant/entity/Application.java
@@ -1,8 +1,8 @@
 package com.unionmate.backend.domain.applicant.entity;
 
 import com.unionmate.backend.domain.applicant.entity.embed.Interview;
-import com.unionmate.backend.domain.recruitment.entity.Recruitment;
-import com.unionmate.backend.domain.recruitment.entity.item.Item;
+import com.unionmate.backend.domain.recruitment.domain.entity.Recruitment;
+import com.unionmate.backend.domain.recruitment.domain.entity.item.Item;
 import com.unionmate.backend.global.entity.BaseEntity;
 import com.unionmate.backend.domain.applicant.entity.embed.Stage;
 import jakarta.persistence.CascadeType;
