+++ src/main/java/com/unionmate/backend/domain/recruitment/domain/entity/item/AnnouncementItem.java
@@ -1,6 +1,6 @@
-package com.unionmate.backend.domain.recruitment.entity.item;
+package com.unionmate.backend.domain.recruitment.domain.entity.item;
 
-import com.unionmate.backend.domain.recruitment.entity.enums.ItemType.DiscriminationValue;
+import com.unionmate.backend.domain.recruitment.domain.entity.enums.ItemType.DiscriminationValue;
 import jakarta.persistence.Column;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
