+++ src/main/java/com/unionmate/backend/domain/recruitment/domain/entity/item/CalendarItem.java
@@ -1,13 +1,14 @@
-package com.unionmate.backend.domain.recruitment.entity.item;
+package com.unionmate.backend.domain.recruitment.domain.entity.item;
 
 import com.unionmate.backend.domain.applicant.entity.column.Answer;
-import com.unionmate.backend.domain.recruitment.entity.enums.ItemType.DiscriminationValue;
-import com.unionmate.backend.global.converter.StringAnswerConverter;
+import com.unionmate.backend.domain.recruitment.domain.entity.enums.ItemType.DiscriminationValue;
+import com.unionmate.backend.global.converter.LocalDateAnswerConverter;
 import jakarta.persistence.Column;
 import jakarta.persistence.Convert;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Lob;
+import java.time.LocalDate;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -19,13 +20,13 @@ import lombok.experimental.SuperBuilder;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-@DiscriminatorValue(DiscriminationValue.TEXT)
-public class TextItem extends Item{
+@DiscriminatorValue(DiscriminationValue.CALENDAR)
+public class CalendarItem extends Item {
 
-  @Column(name = "text", nullable = false, length = 500)
-  private String text;
+  @Column(name = "date", nullable = false)
+  private LocalDate date;
 
-  @Convert(converter = StringAnswerConverter.class)
+  @Convert(converter = LocalDateAnswerConverter.class)
   @Lob
-  private Answer<String> answer;
+  private Answer<LocalDate> answer;
 }
